Add getBigDecimalValue to all dicts. address comments

+++ pinot-core/src/main/java/org/apache/pinot/core/operator/blocks/IntermediateResultsBlock.java
@@ -20,6 +20,7 @@ package org.apache.pinot.core.operator.blocks;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
@@ -361,7 +362,7 @@ public class IntermediateResultsBlock implements Block {
         dataTableBuilder.setColumn(columnIndex, (double) value);
         break;
       case BIG_DECIMAL:
-        dataTableBuilder.setColumn(columnIndex, value);
+        dataTableBuilder.setColumn(columnIndex, (BigDecimal) value);
         break;
       case STRING:
         dataTableBuilder.setColumn(columnIndex, (String) value);
